refactored location finder methods into class
