+++ lib/dhl/bcs/v2/receiver.rb
@@ -8,8 +8,7 @@ module Dhl::Bcs::V2
 
     def self.build(name: nil, **attributes)
       communication = Communication.build(attributes)
-      location_class = location_class_finder(attributes)
-      location = location_builder(location_class, attributes)
+      location = Locator.for(attributes)
       new(attributes.merge(name: name, communication: communication, location: location))
     end
 
@@ -25,24 +24,6 @@ module Dhl::Bcs::V2
         'Communication' => communication.to_soap_hash,
         location.class.name.split('::').last => location.to_soap_hash
       }
-    end
-
-    private
-
-    def self.location_class_finder(attributes)
-      if attributes.key?(:packstation_number)
-        Packstation
-      elsif attributes.key?(:postfilial_number)
-        Postfiliale
-      elsif attributes.key?(:parcel_shop_number)
-        ParcelShop
-      else
-        Address
-      end
-    end
-
-    def self.location_builder(location_class = Address, attributes)
-      location_class.build(attributes)
     end
 
   end
